refactor(merchant acquirer): connect common dto

+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/factory/AuthorizationRequestFactory.java
@@ -3,7 +3,7 @@ package com.processing.merchantacquirer.domain.factory;
 import com.processing.merchantacquirer.domain.StanGenerator;
 import com.processing.merchantacquirer.domain.entity.Merchant;
 import com.processing.merchantacquirer.domain.entity.Terminal;
-import com.processing.merchantacquirer.domain.model.AuthorizationRequest;
+import com.processing.common.dto.authorization.AuthorizationRequest;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -16,7 +16,7 @@ public class AuthorizationRequestFactory {
   private final StanGenerator stanGenerator;
 
   public AuthorizationRequest build(
-          String pan, BigDecimal amount, Terminal terminal, Merchant merchant) {
+          String pan, Integer amount, Terminal terminal, Merchant merchant) {
     LocalDateTime time = LocalDateTime.now();
 
     return AuthorizationRequest.builder()
